chore: refactor canary into a standalone component
